Eliminated LoginFragment and LoginActivity: they only existed because I didn't know a thing about Android development and the code sample with the Button was in an activity :/
