+++ app/src/main/java/io.github/getExposure/profile/ProfileActivity.java
@@ -8,9 +8,15 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import com.facebook.AccessToken;
+import com.facebook.AccessTokenTracker;
+import com.facebook.CallbackManager;
+import com.facebook.FacebookCallback;
+import com.facebook.FacebookException;
 import com.facebook.GraphRequest;
 import com.facebook.GraphResponse;
 import com.facebook.Profile;
+import com.facebook.ProfileTracker;
+import com.facebook.login.LoginResult;
 import com.facebook.login.widget.ProfilePictureView;
 
 import org.json.JSONException;
@@ -18,6 +24,7 @@ import org.json.JSONObject;
 
 import io.github.getExposure.database.DatabaseManager;
 import io.github.getExposure.database.ExposurePhoto;
+import io.github.getExposure.database.ExposureUser;
 import io.github.getExposure.post.PostActivity;
 import io.github.getExposure.R;
 import io.github.getExposure.maps.MapsActivity;
@@ -29,9 +36,47 @@ public class ProfileActivity extends AppCompatActivity {
     private TextView profilename;
     private TextView picsAdded;
     private TextView count;
-    private Button loginViewSwitcher;
+
+    private CallbackManager mCallbackManager;
+
+    // Trackers for the user's profile and access token information
+    private AccessTokenTracker mTokenTracker;
+    private ProfileTracker mProfileTracker;
 
     private DatabaseManager db;
+
+    private FacebookCallback<LoginResult> mFacebookCallback = new FacebookCallback<LoginResult>() {
+        /*
+        Actions to take when the user successfully logs in. Gets an AccessToken and a Profile
+        object for accessing basic profile information. The methods here will execute before
+        methods in getProfileTracker and getAccessTokenTracker
+         */
+        @Override
+        public void onSuccess(LoginResult loginResult) {
+            // The user successfully logs in
+            // AccessToken accessToken = loginResult.getAccessToken();
+        }
+
+        /*
+        Actions to take when the user cancels the login.
+         */
+        @Override
+        public void onCancel() {
+            // The user cancels the login at any point
+            profile = null;
+            // message.setText(LOGGED_OUT_TEXT);
+        }
+
+        /*
+        Actions to take when an error occurs during the login process.
+         */
+        @Override
+        public void onError(FacebookException e) {
+            profile = null;
+            // message.setText(ERROR_TEXT);
+        }
+    };
+
 // Use ID 49-54
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -40,10 +85,15 @@ public class ProfileActivity extends AppCompatActivity {
 
         db = new DatabaseManager(getApplicationContext());
         profile = Profile.getCurrentProfile();
+        mCallbackManager = CallbackManager.Factory.create();
+
+        setupTokenTracker();
+        setupProfileTracker();
+
+        mTokenTracker.startTracking();
+        mProfileTracker.startTracking();
 
         if (profile == null) {
-            Intent loginIntent = new Intent(getApplicationContext(), LoginActivity.class);
-            startActivity(loginIntent);
             return;
         }
 
@@ -55,16 +105,67 @@ public class ProfileActivity extends AppCompatActivity {
 
         picview = (ProfilePictureView) findViewById(R.id.view);
         picview.setProfileId((profile.getId()));
-
-        loginViewSwitcher = (Button) findViewById(R.id.button);
-
-        loginViewSwitcher.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                Intent loginViewIntent = new Intent(getApplicationContext(), LoginActivity.class);
-                startActivity(loginViewIntent);
-            }
-        });
     }
+
+    /*
+    When the activity is resumed, get whatever Profile is currently logged in
+     */
+    @Override
+    public void onResume() {
+        super.onResume();
+        profile = Profile.getCurrentProfile();
+    }
+
+    /*
+    When the activity is closed, stop tracking the profile information
+     */
+    @Override
+    public void onStop() {
+        super.onStop();
+        mProfileTracker.stopTracking();
+        mTokenTracker.stopTracking();
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        mCallbackManager.onActivityResult(requestCode, resultCode, data);
+    }
+
+    private ExposureUser getExposureUser() {
+        long id = Long.parseLong(profile.getId());
+        ExposureUser user = db.getUser(id);
+        if (user == null) {
+            user = new ExposureUser(id, profile.getName(), profile.getProfilePictureUri(100, 100).toString(), "");
+
+            if (!db.insert(user))
+                System.out.println("Something went horribly wrong!!!!");
+        }
+        return user;
+    }
+
+    private void setupTokenTracker() {
+        mTokenTracker = new AccessTokenTracker() {
+            @Override
+            protected void onCurrentAccessTokenChanged(AccessToken oldAccessToken, AccessToken currentAccessToken) {
+
+            }
+        };
+    }
+
+    /*
+    Setup what happens when a new profile is logged in
+     */
+    private void setupProfileTracker() {
+        mProfileTracker = new ProfileTracker() {
+            @Override
+            protected void onCurrentProfileChanged(Profile oldProfile, Profile currentProfile) {
+                profile = Profile.getCurrentProfile();
+                getExposureUser();
+            }
+        };
+    }
+
     private void setCity() {
         profilecity = (TextView) findViewById(R.id.profilecity);
         /* make the API call */
