+++ app/src/main/res/layout/activity_profile_view.xml
@@ -9,6 +9,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textAppearance="?android:attr/textAppearanceMedium"
+        android:text="No Name Found"
         android:id="@+id/profilename"
         android:layout_marginTop="153dp"
         android:textColor="#000000"
@@ -74,11 +75,11 @@
             android:textSize="32dp"
             android:layout_marginBottom="45dp" />
 
-        <Button
+        <com.facebook.login.widget.LoginButton
+            android:id="@+id/button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Logout Screen"
-            android:id="@+id/button"
             android:layout_alignBottom="@+id/layout"
             android:layout_centerHorizontal="true"
             android:layout_gravity="center_horizontal|bottom" />
